+++ src/app/components/product/product.component.ts
@@ -2,20 +2,21 @@ import { Component, Input, Output, EventEmitter, signal } from '@angular/core';
 import { MatCardModule } from '@angular/material/card';
 import { MatDialog } from '@angular/material/dialog';
 import { MatExpansionModule } from '@angular/material/expansion';
-import { ShiftDialogComponent } from '../shift-dialog/shift-dialog.component';
+import { ProductDialogComponent } from '../product-dialog/product-dialog.component';
 import { CommonModule } from '@angular/common';
-import { WcProduct } from './shift.model';
+import { WcProduct } from './product.model';
+import { MatTooltipModule } from '@angular/material/tooltip';
 
 @Component({
-  selector: 'shift',
-  templateUrl: './shift.component.html',
-  styleUrls: ['./shift.component.scss'],
-  imports: [MatCardModule, MatExpansionModule, CommonModule],
+  selector: 'product',
+  templateUrl: './product.component.html',
+  styleUrls: ['./product.component.scss'],
+  imports: [MatCardModule, MatExpansionModule, CommonModule, MatTooltipModule],
 })
-export class ShiftComponent {
-  @Input() shift!: WcProduct;
+export class ProductComponent {
+  @Input() product!: WcProduct;
   @Input() isSelected: boolean = false;
-  @Output() shiftSelected = new EventEmitter<any>();
+  @Output() productSelected = new EventEmitter<any>();
   isSelectedSignal = signal(this.isSelected);
   constructor(private readonly dialog: MatDialog) {}
 
@@ -28,14 +29,14 @@ export class ShiftComponent {
     this.isSelectedSignal.set(this.isSelected);
   }
 
-  onShiftSelect(event: Event): void {
-    this.shiftSelected.emit();
+  onProductSelect(event: Event): void {
+    this.productSelected.emit();
   }
 
   openDetailsDialog(event: Event) {
     event.stopPropagation();
-    this.dialog.open(ShiftDialogComponent, {
-      data: this.shift,
+    this.dialog.open(ProductDialogComponent, {
+      data: this.product,
       width: '800px',
     });
   }