+++ src/app/services/auth.service.ts
@@ -1,8 +1,9 @@
 import { Injectable } from '@angular/core';
-import { BACKEND } from '../../config/http.config';
+import { BACKEND, CREDENTIALS } from '../../config/http.config';
 import { HttpClient, HttpHeaders } from '@angular/common/http';
 import { catchError, map, Observable, of } from 'rxjs';
 import { GuestAuth } from './auth.model';
+import shajs from 'sha.js';
 
 @Injectable({
   providedIn: 'root',
@@ -10,15 +11,43 @@ import { GuestAuth } from './auth.model';
 export class AuthService {
   private readonly customBackend = BACKEND + '/custom/v1';
   private readonly jwtAuthBackend = BACKEND + '/jwt-auth/v1';
-  // headers: HttpHeaders;
+
+  private readonly salt: string;
+  private readonly storedPasswordHash: string;
 
   constructor(private readonly http: HttpClient) {
-    // this.headers = new HttpHeaders().set('Content-Type', 'application/json');
+    this.salt = CREDENTIALS.salt!;
+    this.storedPasswordHash = this.hashPassword(CREDENTIALS.password!);
   }
 
-  private isGuestAuth(obj: any): obj is GuestAuth {
-    return obj && typeof obj.token === 'string';
+  get getStoredPwHash() {
+    return this.storedPasswordHash;
   }
+
+  private hashPassword(password: string): string {
+    return shajs('sha256')
+      .update(password + this.salt)
+      .digest('hex');
+  }
+
+  loginGuestWithPw(passwordInput: string): boolean {
+    const inputHash = this.hashPassword(passwordInput);
+    if (inputHash === this.storedPasswordHash) {
+      localStorage.setItem('auth', inputHash);
+      return true;
+    }
+    return false;
+  }
+
+  isAuthenticated(): boolean {
+    return localStorage.getItem('auth') === this.storedPasswordHash;
+  }
+
+  logout(): void {
+    localStorage.removeItem('auth');
+  }
+
+  // TODO: Maybe useful later
 
   sessionHasNonce(): boolean {
     const nonce = sessionStorage.getItem('nonce');
@@ -36,7 +65,6 @@ export class AuthService {
     );
   }
 
-  // TODO: Maybe useful later
   validateToken(): Observable<boolean> {
     const token = sessionStorage.getItem('token');
     if (!token) {