+++ src/config/http.config.ts
@@ -3,3 +3,4 @@ import { environment } from '../environments/environment';
 export const BACKEND = environment.backendUrl;
 export const CONSUMER_KEY = environment.consumerKey;
 export const CONSUMER_SECRET = environment.consumerSecret;
+export const CREDENTIALS = environment.credentials;