+++ webpack.config.ts
@@ -2,6 +2,11 @@ import Dotenv from 'dotenv-webpack';
 
 module.exports = {
   plugins: [new Dotenv()],
+  resolve: {
+    fallback: {
+      stream: require.resolve('stream-browserify'),
+    },
+  },
 };
 
 /**